added tests, fixed services and storages

+++ src/main/java/ru/yandex/practicum/filmorate/service/UserService.java
@@ -2,9 +2,9 @@ package ru.yandex.practicum.filmorate.service;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
+import ru.yandex.practicum.filmorate.exceptions.DataNotFoundException;
 import ru.yandex.practicum.filmorate.exceptions.ValidationException;
 import ru.yandex.practicum.filmorate.model.User;
-import ru.yandex.practicum.filmorate.storage.InMemoryUserStorage;
 import ru.yandex.practicum.filmorate.storage.UserStorage;
 
 import java.time.LocalDate;
@@ -13,8 +13,6 @@ import java.util.List;
 @Service
 @Slf4j
 public class UserService {
-
-    private static Long nextId = 0L;
     private final UserStorage userStorage;
 
     public UserService(UserStorage userStorage) {
@@ -22,24 +20,27 @@ public class UserService {
     }
 
     public User addToUsers(User user) {
-        userValidation(user);
+        birthdayValidation(user);
         if (user.getName().equals("")) {
             user.setName(user.getLogin());
-        }
-        if(user.getId() == null) {
-            user.setId(genId());
         }
         log.debug("Пользователь {} добавлен", user);
         return userStorage.addToUsers(user);
     }
 
     public User putToUser(User user) {
-        userValidation(user);
-        if (user.getName() == null) {
+        birthdayValidation(user);
+        if (user.getName() == null || user.getName().equals("")) {
             user.setName(user.getLogin());
         }
-        log.debug("Пользователь {} обновлен", user);
-        return userStorage.putToUser(user);
+
+        if (userStorage.getUsers().contains(user)) {
+            userStorage.putToUser(user);
+            log.debug("Пользователь {} обновлен", user);
+            return user;
+        } else {
+            throw new DataNotFoundException("Данного пользователя нет в записях");
+        }
     }
 
     public List<User> getUsers() {
@@ -58,13 +59,17 @@ public class UserService {
     }
 
     public void addFriend(Long id, Long friendId) {
-        log.debug("К пользователю с id: {} в друзья добавлен пользователь с id: {}", id, friendId);
-        userStorage.addFriend(id, friendId);
+        if (userStorage.getUser(id) != null && userStorage.getUser(friendId) != null) {
+            log.debug("К пользователю с id: {} в друзья добавлен пользователь с id: {}", id, friendId);
+            userStorage.addFriend(id, friendId);
+        }
     }
 
-    public User deleteFriend(Long id, Long friendId) {
-        log.debug("У пользователя с id: {} из друзей удален пользователь с id: {}", id, friendId);
-        return userStorage.deleteFriend(id, friendId);
+    public void deleteFriend(Long id, Long friendId) {
+        if (userStorage.getUser(id) != null && userStorage.getUser(friendId) != null) {
+            log.debug("У пользователя с id: {} из друзей удален пользователь с id: {}", id, friendId);
+            userStorage.deleteFriend(id, friendId);
+        }
     }
 
     public List<User> getCommonFriends(Long id, Long otherId) {
@@ -72,40 +77,14 @@ public class UserService {
         return userStorage.getCommonFriends(id, otherId);
     }
 
-    private Long genId() {
-        nextId++;
-        return nextId;
-    }
-
-    private void userValidation(User user) {
-        emailValidation(user);
-        loginValidation(user);
-        birthdayValidation(user);
-    }
-
-    private void emailValidation(User user) {
-        if (user.getEmail() == null) {
-            log.debug("Пользователь {} не прошел валидацию. Полные данные: {}", user.getLogin(), user);
-            throwExceptionType("email");
-        }
-    }
-
-    private void loginValidation(User user) {
-        if (user.getLogin() == null || user.getLogin().contains(" ")) {
-            log.debug("Пользователь {} не прошел валидацию. Полные данные: {}", user.getLogin(), user);
-            throwExceptionType("login");
-        }
-    }
-
     private void birthdayValidation(User user) {
         if (user.getBirthday() == null || user.getBirthday().isAfter(LocalDate.now())) {
             log.debug("Пользователь {} не прошел валидацию. Полные данные: {}", user.getLogin(), user);
-            throwExceptionType("birthday");
+            throw new ValidationException("Некорректный формат поля birthday");
         }
     }
 
-    private void throwExceptionType(String type) {
-        throw new ValidationException("Некорректный формат поля " + type);
+    public void deleteUser(Long id) {
+        userStorage.deleteUser(id);
     }
-
 }
